+++ cms/src/pages/post/Post.tsx
@@ -2,11 +2,14 @@ import { useLoaderData, useLocation } from "react-router";
 import NavAnchor from "../../components/navAnchor/NavAnchor";
 import { Post as PostData } from "../../interfaces/blog";
 import { useUserData } from "../../hooks/useUser";
+import { useModalContext } from "../../hooks/useModal";
+import ConfirmPostDelete from "../../features/posts/components/ConfirmPostDelete";
 
 const Post = () => {
   const location = useLocation();
   const { status, user } = useUserData();
   const currentPost = useLoaderData<PostData>();
+  const { openModal } = useModalContext();
   console.log("currentPost:", currentPost);
 
   // const { postTitle } = useParams<PostParams>();
@@ -26,6 +29,11 @@ const Post = () => {
     <section>
       <header>
         <NavAnchor pathname={pathname} textContent="edit" />
+        <button
+          onClick={() => openModal(<ConfirmPostDelete data={currentPost} />)}
+        >
+          Delete
+        </button>
       </header>
       <article>
         <header>